Make stdout available on error
When the process exits with an error, include the stdout output as well as the error. This makes debugging script failures more straightforward since the rest of the output is available.

For example:

print('hello world')
raise Exception('goodbye world')

callback() will now receive both "goodbye world" error, and the "hello world" output.